some styling, options as objects!

+++ src/main/java/nl/kolkos/cryptoManager/Portfolio.java
@@ -9,15 +9,15 @@ import javax.persistence.Id;
 public class Portfolio {
 	@Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 	
 	private String name;
 	private String description;
 	
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
@@ -33,6 +33,5 @@ public class Portfolio {
 		this.description = description;
 	}
 
-	
-	
+
 }
